Rename list.js to index.js to when user access load the job list
