+++ frontend/.env.example
@@ -1,1 +1,4 @@
 NEXT_PUBLIC_API_URL_NEWSLETTER=<url_api_newsletter>
+NEXT_PUBLIC_API_URL=<url_api>
+API_URL=<url_api>
+NEXT_PUBLIC_LINK_INVITE_DISCORD=<link_invite_discord>
