+++ frontend/package.json
@@ -15,6 +15,7 @@
   "license": "MIT",
   "dependencies": {
     "axios": "^0.27.2",
+    "dayjs": "^1.11.6",
     "next": "^12.3.1",
     "postcss-import": "^14.0.2",
     "react": "^17.0.2",
