+++ frontend/pages/index.js
@@ -1,19 +1,75 @@
+import React, { useState } from "react"
+import axios from "axios"
+import dayjs from "dayjs"
 
-import React from "react"
-import Link from "next/link"
+const Index = (props) => {
+    const [page, setPage] = useState(props.page)
+    const [jobs, setJobs] = useState(props.jobs)
 
-const Index = () => {
+    const loadMoreJobs = async (event) => {
+        event.preventDefault()
+        const nextPage = page + 1
+        setPage(nextPage)
+
+        const response = await axios.get(`${process.env.NEXT_PUBLIC_API_URL}/jobs?page=${nextPage}&totalItens=10`);
+        const newJobs = response.data;
+
+        setJobs([...jobs, ...newJobs])
+    }
     return (
-        <div>                    
-            <div>                
-                <Link href='/'>
-                    <a></a>
-                </Link> 
-                                                          
+        <div>
+            <div>
+                {
+                    jobs.map(job => {
+
+                        return (
+
+                            <div className='grid grid-cols-2 gap-4 m-5 py-3 px-4 border-solid border-2 border-black'>
+                                { job.company &&
+                                    <div className='col-span-2 bg-fuchsia-600 px-1 text-black'>Company: {job.company} <br /></div>
+                                }
+
+                                <div className='bg-fuchsia-600 px-2'>Title: {job.title} </div><br />
+                                { job.platform &&
+                                    <><div className='bg-fuchsia-600 px-1 text-black'>Platform: {job.platform} </div><br /></>
+                                }
+
+                                { job.postedAt &&
+                                    <>
+                                        <div className='bg-fuchsia-600 px-1 text-black'>Post: {dayjs(job.postedAt).format("DD/MM/YYYY HH:mm")} </div><br />
+                                    </>
+                                }
+
+                                <div >
+                                    <a href={job.link} target="_blank" class="bg-blue-500 hover:bg--700 text-white font-bold py-2 px-4 rounded-full">
+                                        See details
+                                        </a>
+                                </div>
+                            </div>
+                        )
+                    })
+                }
+                <br />
+                <div className='text-center'>
+                    <button onClick={(event) => loadMoreJobs(event)} className='bg-blue-500 hover:bg-blue-700  
+                        text-center text-white font-bold py-3 m-5 px-20 rounded-md'>
+                        Click here to load more jobs
+                    </button>
+                </div>
             </div>
-       </div>
-      
+        </div>
     )
+}
+export async function getServerSideProps() {
+    const response = await axios.get(`${process.env.API_URL}/jobs?page=1&totalItens=10`);
+    const jobs = response.data
+
+    return {
+        props: {
+            page: 1,
+            jobs
+        },
+    };
 }
 
 export default Index
